move the icons to the assets folder
