+++ src/assets/icons/SunIcon.jsx
